Add ManagerBase abstraction
